refactor(menus): 메뉴 컨텍스트 패키지 구조 정리
