--- src/main/java/kitchenpos/tobe/menus/domain/MenuGroupRepository.java
@@ -1,7 +1,0 @@
-package kitchenpos.tobe.menus.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
-
-}
