--- src/main/java/kitchenpos/tobe/menus/domain/MenuRepository.java
@@ -1,9 +1,0 @@
-package kitchenpos.tobe.menus.domain;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-
-    List<Menu> findAllByIdIn(final List<Long> menuIds);
-}
