--- src/main/java/kitchenpos/tobe/menus/dto/MenuGroupResponse.java
@@ -1,35 +1,0 @@
-package kitchenpos.tobe.menus.dto;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.tobe.menus.domain.MenuGroup;
-
-public class MenuGroupResponse {
-
-    private final Long id;
-
-    private final String name;
-
-    public MenuGroupResponse(final Long id, final String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static List<MenuGroupResponse> ofList(final List<MenuGroup> menuGroups) {
-        return menuGroups.stream()
-            .map(MenuGroupResponse::of)
-            .collect(Collectors.toList());
-    }
-
-    public static MenuGroupResponse of(final MenuGroup menuGroup) {
-        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
