+++ src/main/java/kitchenpos/tobe/menus/menugroup/domain/MenuGroupRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.tobe.menus.menugroup.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
+}
