+++ src/test/java/kitchenpos/tobe/fixture/MenuProductsFixture.java
@@ -1,8 +1,8 @@
 package kitchenpos.tobe.fixture;
 
 import java.util.Arrays;
-import kitchenpos.tobe.menus.domain.MenuProduct;
-import kitchenpos.tobe.menus.domain.MenuProducts;
+import kitchenpos.tobe.menus.menu.domain.MenuProduct;
+import kitchenpos.tobe.menus.menu.domain.MenuProducts;
 
 public class MenuProductsFixture {
