--- src/test/java/kitchenpos/tobe/menus/domain/FakeMenuValidator.java
@@ -1,10 +1,0 @@
-package kitchenpos.tobe.menus.domain;
-
-import kitchenpos.tobe.common.domain.Validator;
-
-public class FakeMenuValidator implements Validator<Menu> {
-
-    @Override
-    public void validate(final Menu menu) {
-    }
-}
