+++ src/test/java/kitchenpos/tobe/menus/menu/domain/FakeMenuValidator.java
@@ -1,0 +1,10 @@
+package kitchenpos.tobe.menus.menu.domain;
+
+import kitchenpos.tobe.common.domain.Validator;
+
+public class FakeMenuValidator implements Validator<Menu> {
+
+    @Override
+    public void validate(final Menu menu) {
+    }
+}
