+++ src/test/java/kitchenpos/tobe/orders/domain/order/OrderValidatorTest.java
@@ -14,8 +14,8 @@ import kitchenpos.tobe.fixture.MenuProductsFixture;
 import kitchenpos.tobe.fixture.OrderFixture;
 import kitchenpos.tobe.fixture.OrderLineItemFixture;
 import kitchenpos.tobe.fixture.OrderTableFixture;
-import kitchenpos.tobe.menus.domain.FakeMenuValidator;
-import kitchenpos.tobe.menus.domain.MenuRepository;
+import kitchenpos.tobe.menus.menu.domain.FakeMenuValidator;
+import kitchenpos.tobe.menus.menu.domain.MenuRepository;
 import kitchenpos.tobe.orders.domain.ordertable.OrderTable;
 import kitchenpos.tobe.orders.domain.ordertable.OrderTableRepository;
 import org.assertj.core.api.ThrowableAssert;
